--- app/src/main/java/com/techcourse/support/mvc/ManualHandlerAdapter.java
@@ -1,26 +1,0 @@
-package com.techcourse.support.mvc;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerAdapter;
-import webmvc.org.springframework.web.servlet.view.JspView;
-
-public class ManualHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean isSupport(Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(
-            Object handler,
-            HttpServletRequest request,
-            HttpServletResponse response
-    ) throws Exception {
-        String viewName = ((Controller) handler).execute(request, response);
-        return new ModelAndView(new JspView(viewName));
-    }
-}
