+++ app/src/test/java/com/techcourse/support/mvc/HandlerMappingsTest.java
@@ -8,9 +8,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-import webmvc.org.springframework.web.servlet.mvc.tobe.AnnotationHandlerMapping;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerExecution;
+import webmvc.org.springframework.web.servlet.mvc.AnnotationHandlerMapping;
+import webmvc.org.springframework.web.servlet.mvc.HandlerExecution;
 
 class HandlerMappingsTest {
 
@@ -20,27 +19,13 @@ class HandlerMappingsTest {
     @BeforeEach
     void setUp() {
         handlerMappings = new HandlerMappings();
-        handlerMappings.addHandlerMapping(new ManualHandlerMapping());
         handlerMappings.addHandlerMapping(new AnnotationHandlerMapping("com.techcourse.controller"));
         handlerMappings.initialize();
         request = mock(HttpServletRequest.class);
     }
 
     @ParameterizedTest
-    @CsvSource({"/logout", "/register/view", "/register"})
-    void getHandlerFromManualHandlerMapping(String requestURI) {
-        // given
-        when(request.getRequestURI()).thenReturn(requestURI);
-
-        // when
-        Object handler = handlerMappings.getHandler(request);
-
-        // then
-        assertThat(handler).isInstanceOf(Controller.class);
-    }
-
-    @ParameterizedTest
-    @CsvSource({"/login, POST", "/login/view, GET"})
+    @CsvSource({"/logout, GET", "/register/view, GET", "/register, POST", "/login, POST", "/login/view, GET"})
     void getHandlerFromAnnotationHandlerMapping(String requestURI, String method) {
         // given
         when(request.getRequestURI()).thenReturn(requestURI);
